pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
my_robot = "hello_agent.agent_hello:my_first_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
